Continuing with the initial pass the game now starts with UI-derived rules parameters for dimensions, win condition, movement rules, parameters for match-N wins, total games required to win, max games and player types, icon and colors.
The board and its tiles are set up consistent with the requested
dimensions in the match setup. The introduction of the board and tiles are
nicely animated via the physics system; just dropping each from a set
height.

All major UI elements (new dialog, UI root/canvas, main HUD) are animated
with the intent of placing the responsibility of timing everything on an
artist-type in the future. The animations are simple alpha fades but could
be more elaborate move/stretch/squash animations as well.

The final third of the initial check in phase is to get a game to start
and finish, as well as a whole match and loop back the flow to the initial
screen with the New Match dialog.